fix: include dock site module in build
