+++ CMakeLists.txt
@@ -15,25 +15,24 @@ if(MINGW)
 endif()
 
 set(C_SOURCES
-    src/bresenham.c
     src/brush.c
     src/canvas.c
     src/color_context.c
     src/debug.c
     src/dockhost.c
+    src/dock_manager.c
+    src/dock_site.c
     src/document.c
     src/file_open.c
     src/new.c
     src/option_bar.c
     src/palette.c
     src/panitent.c
-    src/primitives_context.c
     src/settings.c
     src/swatch.c
     src/swatch2.c
     src/toolbox.c
     src/viewport.c
-    src/wu_primitives.c
     src/crefptr.c
     src/wic.c
     src/history.c
@@ -54,7 +53,6 @@ if(MSVC)
   set(SOURCE_EXE ${SOURCE_EXE} res/panitent.exe.manifest)
 
   set(HEADERS
-      src/bresenham.h
       src/brush.h
       src/brush_context.h
       src/canvas.h
@@ -65,6 +63,7 @@ if(MSVC)
       src/discordsdk.h
       src/dockhost.h
       src/dock_guides.h
+      src/dock_site.h
       src/document.h
       src/file_open.h
       src/flexible.h
@@ -75,7 +74,6 @@ if(MSVC)
       src/pencil_tool.h
       src/pentablet.h
       src/precomp.h
-      src/primitives_context.h
       src/resource.h
       src/settings.h
       src/settings_dialog.h
@@ -87,7 +85,6 @@ if(MSVC)
       src/util.h
       src/viewport.h
       src/wic.h
-      src/wu_primitives.h
       )
 
   source_group("Headers" FILES ${HEADERS})
